Use one form for both adding/editing collaborators, with conditional email field/validation
