+++ static/js/components/SiteCollaboratorAddPanel.test.tsx
@@ -68,7 +68,7 @@ describe("SiteCollaboratorAddPanel", () => {
 
   it("renders a form with the right props", async () => {
     const { wrapper } = await render()
-    const form = wrapper.find("SiteCollaboratorAddForm")
+    const form = wrapper.find("SiteCollaboratorForm")
     expect(form.exists()).toBe(true)
     expect(form.prop("onSubmit")).toBeDefined()
   })
@@ -81,7 +81,7 @@ describe("SiteCollaboratorAddPanel", () => {
         status: 201
       })
     const { wrapper } = await render()
-    const form = wrapper.find("SiteCollaboratorAddForm")
+    const form = wrapper.find("SiteCollaboratorForm")
     const onSubmit = form.prop("onSubmit")
     await act(async () => {
       // @ts-ignore
@@ -115,7 +115,7 @@ describe("SiteCollaboratorAddPanel", () => {
         status: 400
       })
     const { wrapper } = await render()
-    const form = wrapper.find("SiteCollaboratorAddForm")
+    const form = wrapper.find("SiteCollaboratorForm")
     const onSubmit = form.prop("onSubmit")
     await act(async () => {
       // @ts-ignore
@@ -145,7 +145,7 @@ describe("SiteCollaboratorAddPanel", () => {
         status: 400
       })
     const { wrapper } = await render()
-    const form = wrapper.find("SiteCollaboratorAddForm")
+    const form = wrapper.find("SiteCollaboratorForm")
     const onSubmit = form.prop("onSubmit")
     await act(async () => {
       // @ts-ignore
