+++ static/js/components/SiteCollaboratorAddPanel.tsx
@@ -1,15 +1,18 @@
 import React from "react"
-import { useMutation } from "redux-query-react"
+import { useMutation, useRequest } from "redux-query-react"
 import { RouteComponentProps, useRouteMatch } from "react-router-dom"
 import { FormikHelpers } from "formik"
 import { is } from "ramda"
 
-import SiteCollaboratorAddForm from "./forms/SiteCollaboratorAddForm"
+import SiteCollaboratorForm from "./forms/SiteCollaboratorForm"
 import { siteCollaboratorsUrl } from "../lib/urls"
 import { getResponseBodyError, isErrorResponse } from "../lib/util"
-import { createWebsiteCollaboratorMutation } from "../query-configs/websites"
+import {
+  createWebsiteCollaboratorMutation,
+  websiteCollaboratorsRequest
+} from "../query-configs/websites"
 
-import { WebsiteCollaboratorForm } from "../types/websites"
+import { WebsiteCollaboratorFormData } from "../types/websites"
 
 interface MatchParams {
   username: string
@@ -24,19 +27,23 @@ export default function SiteCollaboratorAddPanel({
   const match = useRouteMatch<MatchParams>()
   const { name } = match.params
 
+  const [{ isPending }] = useRequest(websiteCollaboratorsRequest(name))
   const [collaboratorQueryState, addCollaborator] = useMutation(
     createWebsiteCollaboratorMutation
   )
 
   const onSubmit = async (
-    values: WebsiteCollaboratorForm,
+    values: WebsiteCollaboratorFormData,
     {
       setSubmitting,
       setErrors,
       setStatus
-    }: FormikHelpers<WebsiteCollaboratorForm>
+    }: FormikHelpers<WebsiteCollaboratorFormData>
   ) => {
     if (collaboratorQueryState.isPending) {
+      return
+    }
+    if (isPending) {
       return
     }
     const response = await addCollaborator(name, values)
@@ -68,7 +75,7 @@ export default function SiteCollaboratorAddPanel({
     <div className="narrow-page-body m-3">
       <h3>Add Collaborator</h3>
       <div className="form-container m-3 p-3">
-        <SiteCollaboratorAddForm onSubmit={onSubmit}></SiteCollaboratorAddForm>
+        <SiteCollaboratorForm onSubmit={onSubmit}></SiteCollaboratorForm>
       </div>
     </div>
   )
